Continue refactoring FileUtilities.cpp.
