+++ src/Metropolis/Utilities/FileUtilities.cpp
@@ -270,7 +270,7 @@ bool buildBoxData(Environment* enviro, vector<Molecule>& molecVec, Box* box, SBS
 
 	// If there aren't any molecules, return an error.
   if(box->moleculeCount < 1 || molecVec.size() < 1) {
-  	std::cerr << "The simulation environment should have at least 1 " <<
+  	std::cerr << "The simulation environment should have at least 1 "
 							<< "molecule. (The current count? " << box->moleculeCount
 							<<")." << std::endl;
   	return false;
@@ -582,780 +582,6 @@ bool generatefccBox(Box* box) {
 	return true;
 }
 
-
-
-// ============================================================================
-// ========================= State Scanner ====================================
-// ============================================================================
-
-StateScanner::StateScanner(string filename)
-{
-	universal_filename = filename;
-}
-
-StateScanner::~StateScanner()
-{
-
-}
-
-Environment* StateScanner::readInEnvironment()
-{
-	std::string filename = universal_filename;
-
-    ifstream inFile;
-    inFile.open(filename.c_str());
-    string line;
-    Environment* environment;
-
-    if(inFile.is_open())
-    {
-        getline(inFile, line);
-        environment = getEnvironmentFromLine(line);
-    }
-
-    return environment;
-}
-
-long StateScanner::readInStepNumber()
-{
-    std::string filename = universal_filename;
-
-    long startingStep = 0;
-    ifstream inFile(filename.c_str());
-    string line;
-
-    if (inFile.is_open())
-    {
-        getline(inFile, line); // Environment
-        getline(inFile, line); // Step number
-
-        if (!fromString<long>(line, startingStep))
-            return -1;
-        return startingStep;
-    }
-
-    return -1;
-}
-
-vector<Molecule> StateScanner::readInMolecules()
-{
-	std::string filename = universal_filename;
-
-    vector<Molecule> molecules;
-    ifstream inFile(filename.c_str());
-    string line;
-
-    if(inFile.is_open())
-    {
-        vector<Bond> bonds;
-        vector<Angle> angles;
-        vector<Atom> atoms;
-        vector<Dihedral> dihedrals;
-        vector<Hop> hops;
-
-        //the third line starts the first molecule
-        getline(inFile, line); // envrionment
-        getline(inFile, line); // step number
-        getline(inFile, line); //blank
-
-        Molecule currentMol;
-        int section = 0; // 0 = id, 1 = type, 2 = atom, 3 = bond, 4 = dihedral, 5 = hop, 6 = angle 7 = name
-        int molNum = 0;
-        while(inFile.good())
-        {
-            getline(inFile, line);
-            string hold = line.substr(0, 2);
-            switch(section)
-            {
-                case 0: // id
-                    if(hold.compare("= ") == 0)
-                    {
-                        section++;
-                    }
-                    else
-                    {
-                        currentMol.id = atoi(line.c_str());
-                    }
-                    break;
-                case 1: // type
-                    if(hold.compare("= ") == 0)
-                    {
-                        section++;
-                    }
-                    else
-                    {
-                        currentMol.type = atoi(line.c_str());
-                    }
-                    break;
-                case 2: // atom
-                    if(hold.compare("= ") == 0)
-                    {
-                        section++;
-                    }
-                    else
-                    {
-                       atoms.push_back(getAtomFromLine(line));
-                    }
-                    break;
-                case 3: // bond
-                    if(hold.compare("= ") == 0)
-                    {
-                        section++;
-                    }
-                    else
-                    {
-					   bonds.push_back(getBondFromLine(line));
-                    }
-                    break;
-                case 4: // dihedral
-                    if(hold.compare("= ") == 0)
-                    {
-                        section++;
-                    }
-                    else
-                    {
-                       dihedrals.push_back(getDihedralFromLine(line));
-                    }
-                    break;
-                case 5: // hop
-                    if(hold.compare("= ") == 0)
-                    {
-                        section++;
-                    }
-                    else
-                    {
-                        hops.push_back(getHopFromLine(line));
-                    }
-                    break;
-                case 6: // angle
-                    if(hold.compare("==") == 0)
-                    {
-                        section = 0;
-                        molNum++;
-
-                        //convert all vectors to arrays
-                        Bond *bondArray = (Bond *) malloc(sizeof(Bond) * bonds.size());
-                        Angle *angleArray = (Angle *) malloc(sizeof(Angle) * angles.size());
-                        Atom *atomArray = (Atom *) malloc(sizeof(Atom) * atoms.size());
-                        Dihedral *dihedralArray = (Dihedral *) malloc(sizeof(Dihedral) * dihedrals.size());
-                        Hop *hopArray = (Hop *) malloc(sizeof(Hop) * hops.size());
-
-                        for(int i = 0; i < bonds.size(); i++)
-                        {
-                            bondArray[i] = bonds[i];
-                        }
-						for(int i = 0; i < angles.size(); i++)
-                        {
-                            angleArray[i] = angles[i];
-                        }
-						for(int i = 0; i < atoms.size(); i++)
-                        {
-                            atomArray[i] = atoms[i];
-                        }
-                        for(int i = 0; i < dihedrals.size(); i++)
-                        {
-                            dihedralArray[i] = dihedrals[i];
-                        }
-                        for(int i = 0; i < hops.size(); i++)
-                        {
-                            hopArray[i] = hops[i];
-                        }
-
-                        //assign arrays to molecule
-                        currentMol.atoms = atomArray;
-                        currentMol.numOfAtoms = atoms.size();
-
-                        currentMol.bonds = bondArray;
-                        currentMol.numOfBonds = bonds.size();
-
-                        currentMol.angles = angleArray;
-                        currentMol.numOfAngles = angles.size();
-
-                        currentMol.dihedrals = dihedralArray;
-                        currentMol.numOfDihedrals = dihedrals.size();
-
-                        currentMol.hops = hopArray;
-                        currentMol.numOfHops = hops.size();
-
-                        //add molecule to array of returned molecules
-                        molecules.push_back(currentMol);
-
-                        Molecule newMolec;
-                        currentMol = newMolec;
-
-                        dihedrals.clear();
-                        atoms.clear();
-                        bonds.clear();
-                        angles.clear();
-                        hops.clear();
-                    }
-                    else
-                    {
-                       angles.push_back(getAngleFromLine(line));
-                    }
-                    break;
-            }
-        }
-    }
-    else
-    {
-        return molecules;
-    }
-
-   return molecules;
-}
-
-Angle StateScanner::getAngleFromLine(string line)
-{
-    Angle angle = Angle();
-    char *tokens;
-    char *charLine = (char *) malloc(sizeof(char) * (line.size()+1));
-    strcpy(charLine, line.c_str());
-    tokens = strtok(charLine, " ");
-    int tokenNumber = 0;
-
-    while(tokens != NULL)
-    {
-       switch(tokenNumber)
-       {
-           case 0:
-               angle.atom1 = atoi(tokens);
-               break;
-            case 1:
-               angle.atom2 = atoi(tokens);
-               break;
-            case 2:
-               angle.value = atof(tokens);
-               break;
-            case 3:
-               if(atoi(tokens) == 1)
-                   angle.variable = true;
-               else
-                   angle.variable = false;
-       }
-
-       tokens = strtok(NULL, " ");
-       tokenNumber++;
-    }
-
-    free (charLine);
-
-    return angle;
-}
-
-Atom StateScanner::getAtomFromLine(string line)
-{
-    Atom atom = createAtom(-1,-1,-1,-1,-1,-1);
-    char *tokens;
-    char *charLine = (char *) malloc(sizeof(char) * (line.size()+1));
-    strcpy(charLine, line.c_str());
-    tokens = strtok(charLine, " ");
-    int tokenNumber = 0;
-
-    while(tokens != NULL)
-    {
-        switch(tokenNumber)
-        {
-            case 0:
-                atom.id = atoi(tokens);
-                break;
-            case 1:
-                atom.x = atof(tokens);
-                break;
-            case 2:
-                atom.y = atof(tokens);
-                break;
-            case 3:
-                atom.z = atof(tokens);
-                break;
-            case 4:
-                atom.sigma = atof(tokens);
-                break;
-            case 5:
-                atom.epsilon = atof(tokens);
-                break;
-            case 6:
-                atom.charge = atof(tokens);
-                break;
-            case 7:
-	        *atom.name = std::string(tokens);
- 	        break;
-        }
-
-        tokens = strtok(NULL, " ");
-        tokenNumber++;
-    }
-
-    free (charLine);
-
-    return atom;
-}
-
-Bond StateScanner::getBondFromLine(string line)
-{
-    Bond bond = Bond();
-    char *tokens;
-    char *charLine = (char *) malloc(sizeof(char) * (line.size()+1));
-    strcpy(charLine, line.c_str());
-    tokens = strtok(charLine, " ");
-    int tokenNumber = 0;
-
-    while(tokens != NULL)
-    {
-        switch(tokenNumber)
-        {
-            case 0: // atom 1
-                bond.atom1 = atoi(tokens);
-                break;
-            case 1: // atom 2
-                bond.atom2 = atoi(tokens);
-                break;
-            case 2: // value
-                bond.distance = atof(tokens);
-                break;
-            case 3: // variable
-                if(atoi(tokens) == 1)
-                    bond.variable = true;
-                else
-                    bond.variable = false;
-                break;
-        }
-
-        tokens = strtok(NULL, " ");
-        tokenNumber++;
-    }
-
-    free (charLine);
-
-    return bond;
-}
-
-Dihedral StateScanner::getDihedralFromLine(string line)
-{
-    Dihedral dihedral = Dihedral();
-
-    char *tokens;
-    char *charLine = (char *) malloc(sizeof(char) * (line.size()+1));
-    strcpy(charLine, line.c_str());
-    tokens = strtok(charLine, " ");
-    int tokenNumber = 0;
-    while(tokens != NULL)
-    {
-        switch(tokenNumber)
-        {
-            case 0:
-                dihedral.atom1 = atoi(tokens);
-                break;
-            case 1:
-                dihedral.atom2 = atoi(tokens);
-                break;
-            case 2:
-                dihedral.value = atof(tokens);
-                break;
-            case 3:
-                if(atoi(tokens) == 1)
-                    dihedral.variable = true;
-                else
-                    dihedral.variable = false;
-                break;
-        }
-
-        tokens = strtok(NULL, " ");
-        tokenNumber++;
-    }
-
-    free (charLine);
-
-    return dihedral;
-}
-
-Environment* StateScanner::getEnvironmentFromLine(string line)
-{
-    Environment* environment = new Environment();
-    char *tokens;
-    char *charLine = (char *) malloc(sizeof(char) * (line.size()+1));
-    strcpy(charLine, line.c_str());
-    tokens = strtok(charLine, " ");
-    int numOfAtoms, tokenNumber = 0;
-    Real x,y,z,cutoff;
-
-    while(tokens != NULL)
-    {
-        switch(tokenNumber)
-        {
-            case 0:
-                environment->x = atof(tokens);
-                break;
-            case 1:
-                environment->y = atof(tokens);
-                break;
-            case 2:
-                environment->z = atof(tokens);
-                break;
-            case 3:
-                environment->numOfMolecules = atoi(tokens);
-                break;
-            case 4:
-                environment->numOfAtoms = atoi(tokens);
-                break;
-            case 5:
-                environment->temp = atof(tokens);
-                break;
-            case 6:
-                environment->cutoff = atof(tokens);
-                break;
-            case 7:
-                environment->maxTranslation = atof(tokens);
-                break;
-            case 8:
-                environment->maxRotation = atof(tokens);
-                break;
-            case 9:
-                parsePrimaryIndexDefinitions(environment, std::string(tokens));
-		        environment->randomseed =  atoi(tokens + (strlen(tokens) + 1));
-                break;
-            case 10:
-                environment->randomseed = atoi(tokens);
-                break;
-        }
-
-        tokens = strtok(NULL, " ");
-        tokenNumber++;
-    }
-
-    free (charLine);
-
-    return environment;
-}
-
-void StateScanner::parsePrimaryIndexDefinitions(Environment* enviro, string definitions)
-{
-    *enviro->primaryAtomIndexConfigLine = definitions;
-    char *indexVector;
-    char *charLine = (char *) malloc(sizeof(char) * (definitions.size()+1));
-    strcpy(charLine, definitions.c_str());
-    indexVector = strtok(charLine, ",");
-
-    for (int i = 0; indexVector ; i++)
-    {
-    	std::string sIndexVector = indexVector;
-    	enviro->primaryAtomIndexDefinitions++;
-    	char* c;
-    	if ((c=strchr(indexVector, '['))!=NULL)
-    	{
-    	    indexVector = (indexVector + 1);
-    	    while ((c=strchr(indexVector, ']'))==NULL)
-    	    {
-        		int currentPrimaryIndex = atoi(indexVector);
-        		(*(enviro->primaryAtomIndexArray)).push_back(new std::vector<int>);
-        		(*(*(enviro->primaryAtomIndexArray))[i]).push_back(currentPrimaryIndex - 1);
-        		indexVector = strtok(NULL, ",");
-    	    }
-    	    *c = '\0';
-    	    int currentPrimaryIndex = atoi(indexVector);
-    	    (*(enviro->primaryAtomIndexArray)).push_back(new std::vector<int>);
-    	    (*(*(enviro->primaryAtomIndexArray))[i]).push_back(currentPrimaryIndex - 1);
-    	    indexVector =strtok(NULL, ",");
-    	    continue;
-    	}
-
-        int currentPrimaryIndex = atoi(indexVector);
-    	(*(enviro->primaryAtomIndexArray)).push_back(new std::vector<int>);
-    	(*(*(enviro->primaryAtomIndexArray))[i]).push_back(currentPrimaryIndex - 1);
-    	indexVector = strtok(NULL, ",");
-    }
-    free (charLine);
-}
-
-Hop StateScanner::getHopFromLine(string line)
-{
-    Hop hop = Hop();
-    char *tokens;
-    char *charLine = (char *) malloc(sizeof(char) * (line.size()+1));
-    strcpy(charLine, line.c_str());
-    tokens = strtok(charLine, " ");
-    int tokenNumber = 0;
-
-    while(tokens != NULL)
-    {
-        switch(tokenNumber)
-        {
-            case 0:
-                hop.atom1 = atoi(tokens);
-                break;
-            case 1:
-                hop.atom2 = atoi(tokens);
-                break;
-            case 2:
-                hop.hop = atoi(tokens);
-                break;
-        }
-
-        tokens = strtok(NULL, " ");
-        tokenNumber++;
-    }
-
-    free (charLine);
-
-    return hop;
-}
-
-void StateScanner::outputState(Environment *environment, Molecule *molecules, int numOfMolecules, int step, string filename)
-{
-    ofstream outFile;
-    outFile.open(filename.c_str());
-    outFile << environment->x << " " << environment->y << " "
-        << environment->z << " " << environment->numOfMolecules << " "
-        << environment->numOfAtoms << " " << environment->temp << " "
-        << environment->cutoff << " " << environment->maxTranslation << " "
-        << environment->maxRotation << " " << *environment->primaryAtomIndexConfigLine << " "
-        << environment->randomseed << " "
-        << std::endl << step << std::endl << std::endl;
-
-    for(int i=0; i<numOfMolecules; i++)
-    {
-        Molecule currentMol = molecules[i];
-        outFile << currentMol.id << std::endl;
-
-        outFile << "= Type" << std::endl;
-        outFile << currentMol.type << std::endl;
-
-        // Write atoms
-        outFile << "= Atoms" << std::endl;
-        for(int j=0; j<currentMol.numOfAtoms; j++)
-        {
-            Atom currentAtom = currentMol.atoms[j];
-            outFile << currentAtom.id << " "
-                << currentAtom.x << " " << currentAtom.y
-                << " " << currentAtom.z << " "
-                << currentAtom.sigma << " "
-                << currentAtom.epsilon << " "
-                << currentAtom.charge << " "
-		        << *currentAtom.name << " " << std::endl;
-        }
-
-        // Write bonds
-        outFile << "= Bonds" << std::endl;
-        for(int j=0; j<currentMol.numOfBonds; j++)
-        {
-            Bond currentBond = currentMol.bonds[j];
-            outFile << currentBond.atom1 << " "
-                << currentBond.atom2 << " "
-                << currentBond.distance << " ";
-
-            if(currentBond.variable)
-            {
-                outFile << "1" << std::endl;
-            }
-            else
-            {
-                outFile << "0" << std:: endl;
-            }
-        }
-
-        // Write dihedrals
-        outFile << "= Dihedrals" << std::endl;
-        for(int j=0; j<currentMol.numOfDihedrals; j++)
-        {
-            Dihedral currentDi = currentMol.dihedrals[j];
-            outFile << currentDi.atom1 << " "
-                << currentDi.atom2 << " "
-                << currentDi.value << " ";
-
-            if(currentDi.variable)
-            {
-                outFile << "1" << std::endl;
-            }
-            else
-            {
-                outFile << "0" << std::endl;
-            }
-        }
-
-        // Write hops
-        outFile << "= Hops" << std::endl;
-        for(int j=0; j<currentMol.numOfHops; j++)
-        {
-            Hop currentHop = currentMol.hops[j];
-
-            outFile << currentHop.atom1 << " "
-                << currentHop.atom2 << " "
-                << currentHop.hop << std::endl;
-        }
-
-        // Write angles
-        outFile << "= Angles" << std::endl;
-        for(int j=0; j<currentMol.numOfAngles; j++)
-        {
-            Angle currentAngle = currentMol.angles[j];
-
-            outFile << currentAngle.atom1 << " "
-                << currentAngle.atom2 << " "
-                << currentAngle.value << " ";
-
-            if(currentAngle.variable)
-            {
-                outFile << "1" << std::endl;
-            }
-            else
-            {
-                outFile << "0" << std::endl;
-            }
-        }
-
-        outFile << "==" << std::endl;
-    }
-
-    outFile.close();
-}
-
-//=============================================================================
-//======================== SB Scanner - Reads OPLSAA.sb =======================
-//=============================================================================
-void SBScanner::processBond(string line) {
-	string atom1, atom2;
-	if(line.substr(1,1) == " ") {
-		atom1 = line.substr(0,1);
-	} else {
-		atom1 = line.substr(0,2);
-	}
-	if(line.substr(4,1) == " ") {
-		atom2 = line.substr(3,1);
-	} else {
-		atom2 = line.substr(3,2);
-	}
-	double forceK, bondDist;
-	string rest_of_line = line.substr(6, line.size()-6);
-	stringstream ss(rest_of_line);
-	ss >> forceK >> bondDist;
-	bondDataMap[atom1][atom2].kBond = forceK;
-	bondDataMap[atom2][atom1].kBond = forceK;
-	bondDataMap[atom1][atom2].eqBondDist = bondDist;
-	bondDataMap[atom2][atom1].eqBondDist = bondDist;
-}
-
-void SBScanner::processAngle(string line) {
-	string end1 = line.substr(0, 2);
-	string midAtom = line.substr(3, 2);
-	string end2 = line.substr(6, 2);
-	string rest_of_line = line.substr(8, line.size()-8);
-	double angleK, angle;
-	stringstream ss(rest_of_line);
-	ss >> angleK >> angle;
-	angleDataMap[end1][midAtom][end2].kAngle = angleK;
-	angleDataMap[end2][midAtom][end1].kAngle = angleK;
-	angleDataMap[end1][midAtom][end2].eqAngle = angle;
-	angleDataMap[end2][midAtom][end1].eqAngle = angle;
-}
-
-void SBScanner::processLine(string line) {
-	if(line.at(5) == '-')
-		processAngle(line);
-	else
-		processBond(line);
-}
-
-SBScanner::SBScanner() {
-
-}
-
-SBScanner::~SBScanner() {
-	bondDataMap.clear();
-	angleDataMap.clear();
-}
-
-bool SBScanner::readInSB(string filename) {
-	fileName = filename;
-
-	if (filename.empty()) {
-		std::cerr << "Error: readInOpls(): empty filename given" << std::endl;
-		return false;
-	}
-
-	int numOfLines=0;
-    ifstream oplsScanner(filename.c_str());
-    if (!oplsScanner.is_open()) {
-    	std::cerr << "Error: readInOpls(): could not open file (" << filename << ")" << std::endl;
-        return false;
-    } else {
-        string line;
-        while (oplsScanner.good()) {
-            numOfLines++;
-            getline(oplsScanner,line);
-
-            //check if it is a commented line,
-            //or if it is a title lines
-            try {
-                if(line.at(0) != '*' && line.at(0) != '\"' && line.at(0) != '#')
-				{
-					processLine(line);
-                    //cout <<(line) << endl;
-				}
-            } catch (std::out_of_range& e) {
-            	// Eat the exception and continue...why aren't we failing?
-            }
-        }
-        oplsScanner.close();
-    }
-	return true;
-}
-
-map<string, BondData> SBScanner::getBondMap(string atom1) {
-	if(bondDataMap.count(atom1) == 0) {
-		map<string, BondData> blankMap;
-		return blankMap;
-	} else {
-		return bondDataMap[atom1];
-	}
-}
-
-map<string, map<string, AngleData> > SBScanner::getAngleMap(string endpoint1) {
-	if(angleDataMap.count(endpoint1) == 0) {
-		map<string, map<string, AngleData> > blankMap;
-		return blankMap;
-	} else {
-		return angleDataMap[endpoint1];
-	}
-}
-
-
-Real SBScanner::getKBond(string atom1, string atom2) {
-	if(bondDataMap.count(atom1) == 0)
-		return -1;
-	if(bondDataMap[atom1].count(atom2) == 0)
-		return -1;
-	return bondDataMap[atom1][atom2].kBond;
-}
-
-Real SBScanner::getEqBondDist(string atom1, string atom2) {
-	if(bondDataMap.count(atom1) == 0)
-		return -1;
-	if(bondDataMap[atom1].count(atom2) == 0)
-		return -1;
-	return bondDataMap[atom1][atom2].eqBondDist;
-}
-
-Real SBScanner::getKAngle(string endpoint1, string middleAtom, string endpoint2) {
-	if(angleDataMap.count(endpoint1) == 0)
-		return -1;
-	if(angleDataMap[endpoint1].count(middleAtom) == 0)
-		return -1;
-	if(angleDataMap[endpoint1][middleAtom].count(endpoint2) == 0)
-		return -1;
-	return angleDataMap[endpoint1][middleAtom][endpoint2].kAngle;
-}
-
-Real SBScanner::getEqAngle(string endpoint1, string middleAtom, string endpoint2) {
-	if(angleDataMap.count(endpoint1) == 0)
-		return -1;
-	if(angleDataMap[endpoint1].count(middleAtom) == 0)
-		return -1;
-	if(angleDataMap[endpoint1][middleAtom].count(endpoint2) == 0)
-		return -1;
-	return angleDataMap[endpoint1][middleAtom][endpoint2].eqAngle;
-}
-
-
 // ============================================================================
 // ======================= Logging Functions ==================================
 // ============================================================================
@@ -1418,18 +644,18 @@ void writeToLog(stringstream& ss, int stamp) {
 	ss.clear();
 }
 
-static inline std::string& rtrim(std::string& s) {
+std::string& rtrim(std::string& s) {
 	s.erase(s.find_last_not_of(TRIMMED_CHARS) + 1);
   return s;
 }
 
-static inline std::string& ltrim(std::string& s) {
+std::string& ltrim(std::string& s) {
   std::size_t len = s.find_first_not_of(TRIMMED_CHARS);
   if (len != std::string::npos)
     s.erase(0, len);
   return s;
 }
 
-static inline std::string& trim(std::string& s) {
+std::string& trim(std::string& s) {
   return ltrim(rtrim(s));
 }
